Finished but barely. I had a very tough time with keylisteners.
I could not get the string to be displayed. I managed to get the input
from the console and you would be able to see if you played. The storing
and grabbing info from the file works as expected. Some weird bugs exist
but no time to fix. One bug I have found is just a name not appearing.
Also, some scores may be overwritten if they are the same so i changed
the scoring system so that almost never happens.